Refactor run architecture modules
